Create Create Aurora DB (MySql)
Creation of an Aurora MySQL database. Verify that the SG, NACL, and route tables are all configured to allow communication between an instance in a public subnet and an RDS database in a private subnet. After verifying the appropriate settings are in place, the DB is created. Have an EC2 instance as a bastion to allow us to connect to our private database from an external source.