Create Deploy an Amazon RDS Multi-AZ and Read Replica in AWS
We will enable Multi-AZ and backups, create a read replica, promote a read replica, and update the RDS endpoint in Route 53. Multi-AZ and read replicas serve different purposes with RDS. Multi-AZ is strictly for failover, as the standby instances cannot be read from by an application. Read replicas are used for improved performance and migrations.